Moved fpulling related helper functionalities to a separate file
